refactor: merge client tests location with source and add client test-utils
