+++ src/client/pages/Support/Pages/SupportHome.test.tsx
@@ -4,22 +4,21 @@ import {
 import { BrowserRouter } from 'react-router-dom';
 import { CaseEditInformation } from 'blaise-api-node-client';
 import Organisation from 'blaise-api-node-client/lib/cjs/enums/organisation';
-import UserRole from '../../../../client/types/UserTypes';
-import { getCaseSearchResults, getSurveys } from '../../../../client/api/NodeApi';
+import UserRole from '../../../types/UserTypes';
+import { getCaseSearchResults, getSurveys } from '../../../api/NodeApi';
 import { Survey } from '../../../../common/interfaces/surveyInterface';
-import FilteredSurveyListMockObject from '../../MockObjects/SurveyMockObjects';
-import userMockObject from '../../../server/mockObjects/userMockObject';
-import ResearchHome from '../../../../client/pages/Research/Pages/ResearchHome';
-import { CaseEditInformationListMockObject } from '../../../server/mockObjects/CaseMockObject';
+import FilteredSurveyListMockObject from '../../../test-utils/SurveyMockObjects';
+import userMockObject from '../../../../server/test-utils/userMockObject';
+import SupportHome from './SupportHome';
+import { CaseEditInformationListMockObject } from '../../../../server/test-utils/CaseMockObject';
 
 // set global vars
-const userRole:string = UserRole.FRS_Researcher;
+const userRole:string = UserRole.Survey_Support;
 let view:RenderResult;
-
 // set mocks
-vi.mock('../../../../client/api/NodeApi', async () => {
-  const actual = await vi.importActual<typeof import('../../../../client/api/NodeApi')>(
-    '../../../../client/api/NodeApi');
+vi.mock('../../../api/NodeApi', async () => {
+  const actual = await vi.importActual<typeof import('../../../api/NodeApi')>(
+    '../../../api/NodeApi');
 
   return {
     ...actual,
@@ -39,7 +38,7 @@ describe('Given there are surveys available in blaise', () => {
     getCaseInformationMock.mockImplementation(() => Promise.resolve(CaseEditInformationListMockObject));
   });
 
-  it('should render the Research page correctly when surveys are returned', async () => {
+  it('should render the Support page correctly when surveys are returned', async () => {
     // arrange
     const user = userMockObject;
     user.role = userRole;
@@ -48,14 +47,14 @@ describe('Given there are surveys available in blaise', () => {
     await act(async () => {
       view = render(
         <BrowserRouter>
-          <ResearchHome user={user} />
+          <SupportHome user={user} />
         </BrowserRouter>,
       );
     });
 
     // assert
     expect(view).toMatchSnapshot(
-      'ResearchPageSurveysReturned',
+      'SupportPageSurveysReturned',
     );
   });
 
@@ -68,7 +67,7 @@ describe('Given there are surveys available in blaise', () => {
     await act(async () => {
       view = render(
         <BrowserRouter>
-          <ResearchHome user={user} />
+          <SupportHome user={user} />
         </BrowserRouter>,
       );
     });
@@ -88,13 +87,13 @@ describe('Given there are surveys available in blaise', () => {
 
       expect(questionnaireListView).toHaveTextContent(defaultQuestionnaire.questionnaireName.replace('_EDIT', ''));
 
-      const questionnaireView = view.getByTestId(`${defaultQuestionnaireName}-Research-Content`);
+      const questionnaireView = view.getByTestId(`${defaultQuestionnaireName}-Support-Content`);
       expect(questionnaireView).toHaveTextContent(String(defaultQuestionnaire.fieldPeriod));
       expect(questionnaireView).toHaveTextContent(String(defaultQuestionnaire.numberOfCases));
     });
   });
 
-  it('should render the Research page correctly and Search button disabled because search text not entered', async () => {
+  it('should render the Support page correctly and Search button disabled because search text not entered', async () => {
     // arrange
     const user = userMockObject;
     user.role = userRole;
@@ -102,7 +101,7 @@ describe('Given there are surveys available in blaise', () => {
     await act(async () => {
       view = render(
         <BrowserRouter>
-          <ResearchHome user={user} />
+          <SupportHome user={user} />
         </BrowserRouter>,
       );
     });
@@ -112,11 +111,11 @@ describe('Given there are surveys available in blaise', () => {
 
     // assert
     expect(view).toMatchSnapshot(
-      'ResearchPageSurveysReturnedSearchInitial',
+      'SupportPageSurveysReturnedSearchInitial',
     );
   });
 
-  it('should render the Research page correctly and Search button enabled after search text entered', async () => {
+  it('should render the Support page correctly and Search button enabled after search text entered', async () => {
     // arrange
     const user = userMockObject;
     user.role = userRole;
@@ -124,7 +123,7 @@ describe('Given there are surveys available in blaise', () => {
     await act(async () => {
       view = render(
         <BrowserRouter>
-          <ResearchHome user={user} />
+          <SupportHome user={user} />
         </BrowserRouter>,
       );
     });
@@ -136,11 +135,11 @@ describe('Given there are surveys available in blaise', () => {
 
     // assert
     expect(view).toMatchSnapshot(
-      'ResearchPageSurveysReturnedSearchTextEntered',
+      'SupportPageSurveysReturnedSearchTextEntered',
     );
   });
 
-  it('should render the Research page correctly when surveys are returned and search used', async () => {
+  it('should render the Support page correctly when surveys are returned and search used', async () => {
     // arrange
     const user = userMockObject;
     user.role = userRole;
@@ -148,7 +147,7 @@ describe('Given there are surveys available in blaise', () => {
     await act(async () => {
       view = render(
         <BrowserRouter>
-          <ResearchHome user={user} />
+          <SupportHome user={user} />
         </BrowserRouter>,
       );
     });
@@ -161,7 +160,7 @@ describe('Given there are surveys available in blaise', () => {
 
     // assert
     expect(view).toMatchSnapshot(
-      'ResearchPageSurveysReturnedSearchUsed',
+      'SupportPageSurveysReturnedSearchUsed',
     );
   });
 
@@ -173,7 +172,7 @@ describe('Given there are surveys available in blaise', () => {
     await act(async () => {
       view = render(
         <BrowserRouter>
-          <ResearchHome user={user} />
+          <SupportHome user={user} />
         </BrowserRouter>,
       );
     });
@@ -200,7 +199,7 @@ describe('Given there are surveys available in blaise', () => {
 
       expect(questionnaireListView).toHaveTextContent(defaultQuestionnaire.questionnaireName.replace('_EDIT', ''));
 
-      const questionnaireView = view.getByTestId(`${defaultQuestionnaireName}-Research-Content`);
+      const questionnaireView = view.getByTestId(`${defaultQuestionnaireName}-Support-Content`);
       expect(questionnaireView).toHaveTextContent(String(defaultQuestionnaire.fieldPeriod));
       expect(questionnaireView).toHaveTextContent(String(defaultQuestionnaire.numberOfCases));
 
@@ -215,7 +214,7 @@ describe('Given there are surveys available in blaise', () => {
         expect(outcomeRows[index]).toHaveTextContent(caseDetails.outcome.toString());
         expect(interviewerRows[index]).toHaveTextContent(caseDetails.interviewer);
         expect(organisationRows[index]).toHaveTextContent(Organisation[caseDetails.organisation]?.toString() ?? '');
-        expect(linksRows[index]).toHaveTextContent('Edit case | View case');
+        expect(linksRows[index]).toHaveTextContent('Edit interviewer case | View interviewer case');
       });
     });
   });
@@ -239,14 +238,14 @@ describe('Given there are no surveys available in blaise', () => {
     await act(async () => {
       view = render(
         <BrowserRouter>
-          <ResearchHome user={user} />
+          <SupportHome user={user} />
         </BrowserRouter>,
       );
     });
 
     // assert
     expect(view).toMatchSnapshot(
-      'ResearchPageNoSurveysReturned',
+      'SupportPageNoSurveysReturned',
     );
   });
 
@@ -259,7 +258,7 @@ describe('Given there are no surveys available in blaise', () => {
     await act(async () => {
       view = render(
         <BrowserRouter>
-          <ResearchHome user={user} />
+          <SupportHome user={user} />
         </BrowserRouter>,
       );
     });
@@ -288,7 +287,7 @@ describe('Given there the blaise rest api is not available', () => {
     await act(async () => {
       view = render(
         <BrowserRouter>
-          <ResearchHome user={user} />
+          <SupportHome user={user} />
         </BrowserRouter>,
       );
     });
@@ -307,14 +306,14 @@ describe('Given there the blaise rest api is not available', () => {
     await act(async () => {
       view = render(
         <BrowserRouter>
-          <ResearchHome user={user} />
+          <SupportHome user={user} />
         </BrowserRouter>,
       );
     });
 
     // assert
     expect(view).toMatchSnapshot(
-      'ResearchPageError',
+      'SupportPageError',
     );
   });
 });
@@ -336,7 +335,7 @@ describe('Given there is an error that triggered a catch all 404 or 500 response
     user.role = userRole;
 
     // act
-    render(<ResearchHome user={user} />);
+    render(<SupportHome user={user} />);
 
     // assert
     expect(await screen.findByTestId('ErrorMessage')).toHaveTextContent('Something went wrong');
