+++ src/client/utils/supervisorInformationMapper.test.ts
@@ -1,7 +1,7 @@
 import { CaseEditInformation, CaseOutcome, EditedStatus } from 'blaise-api-node-client';
 import Organisation from 'blaise-api-node-client/lib/cjs/enums/organisation';
-import mapSupervisorInformaiton from '../../../client/utils/supervisorInformationMapper';
-import { SupervisorInformation } from '../../../client/types/supervisorInterface';
+import mapSupervisorInformaiton from './supervisorInformationMapper';
+import { SupervisorInformation } from '../types/supervisorInterface';
 
 describe('Map editor informaiton', () => {
   const editors = [{
